requirements.txt
pycryptodome
dbus-python
gatt
matplotlib
numpy
opencv-python
mediapipe
skicit-learn
tensorflow
